# Participium/api/openapi.yaml
openapi: 3.0.4
info:
  title: Participium - OpenAPI 1.0
  description: Participium Server based.
  termsOfService: https://swagger.io/terms/
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.12
...
paths:
...
  /reports:
    post:
      tags:
        - Reports
      summary: User upload a new Report
      operationId: uploadReport
      requestBody:
        description: Created user object
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Report"
          application/xml:
            schema:
              $ref: "#/components/schemas/Report"
          application/x-www-form-urlencoded:
            schema:
              $ref: "#/components/schemas/Report"
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Report"
            application/xml:
              schema:
                $ref: "#/components/schemas/Report"
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    get:
      tags:
        - Reports
      summary: Retrieve all public Reports for map visualization
      description: Get all approved Reports to display on the public map
      operationId: getReports
      parameters: []
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Report"
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /reports/stats:
    get:
      tags:
        - Reports
      summary: Get reports statistics
      description: Retrieve various statistics about reports in the system. The filters can be applied using query parameters.
        Available filters include date range (`fromDate`, `toDate`), predefined periods (`period`), and report categories (`category`).
        For multiple categories use multiple `category` parameters, e.g., `category=sanification&category=infrastructure`.
      parameters:
        - in: query
          name: fromDate
          schema:
            type: string
            format: date-time
          required: false
          description: Start date for filtering reports (inclusive)
        - in: query
          name: toDate
          schema:
            type: string
            format: date-time
          required: false
          description: End date for filtering reports (inclusive)
        - in: query
          name: period
          schema:
            type: string
            enum: [daily, weekly, monthly, yearly]
          required: false
          description: Predefined period for filtering reports
        - in: query
          name: category
          schema:
              $ref: "#/components/schemas/OfficeType"
          required: false
          description: Filter reports by their state
          
      operationId: getReportStats
      
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Statistics"
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /reports/{id}/message:
    post:
      tags:
        - Reports
      summary: Officer sends a message to report author
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                message:
                  type: string
                  example: "Aggiornamento sul tuo report"
              required:
                - message
      responses:
        "201":
          description: successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notification"
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /internal-messages/report/{reportId}:
    get:
      tags:
        - Internal Messages
      summary: List internal messages for a report
      parameters:
        - in: path
          name: reportId
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/InternalMessage"
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /reports/{id}/followers:
    get:
      tags:
        - Reports
      summary: List users following a report
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /reports/{id}/follow:
    post:
      tags:
        - Reports
      summary: Follow a report
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "201":
          description: follow relation created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Follow"
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      tags:
        - Reports
      summary: Unfollow a report
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: follow relation deleted
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /telegram/reports:
    get:
      tags:
        - Telegram
      summary: Get reports for Telegram notifications
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Report"
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    post:
      tags:
        - Telegram
      summary: Update reports for Telegram notifications
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                reportIds:
                  type: array
                  items:
                    type: integer
              required:
                - reportIds
      responses:
        "200":
          description: successful operation
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /telegram/reports/{reportID}:
    post:
      tags:
        - Telegram
      summary: Add a report to Telegram notifications
      parameters:
        - in: path
          name: reportID
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: successful operation
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      tags:
        - Telegram
      summary: Remove a report from Telegram notifications
      parameters:
        - in: path
          name: reportID
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: successful operation
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
components:
  schemas:
    OfficeType:
      type: string
      enum: [ "infrastructure", "environment","safety", "sanitation","transport","organization","other"]
    OfficerRole:
      type: string
      enum: ["municipal_public_relations_officer", "municipal_administrator","technical_office_staff"]
...
    User:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 2
        username:
          type: string
          example: username
        firstName:
          type: string
          example: Mario
        lastName:
          type: string
          example: Rossi
        password:
          type: string
          example: placeholder
        email:
          type: string
          example: mail@domain.com
        avatar:
          type: string
          example: /uploads/avatars/167123123-avatar.jpg
        telegramUsername:
          type: string
          example: my_telegram
        emailNotifications:
          type: boolean
          example: true
        followedReports:
          type: array
          items:
            $ref: "#/components/schemas/Report"
      xml:
        name: user
...
    Location:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        Coordinates:
          type: object
          properties:
            longitude:
              type: number
            latitude:
              type: number
    Report:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        location:
          $ref: "#/components/schemas/Location"
        Author:
          $ref: "#/components/schemas/User"
        Anonymity:
          type: boolean
        Date:
          type: string
        Category:
          $ref: "#/components/schemas/OfficeType"
        Document:
          type: object
          properties:
            Description:
              type: string
              minLength: 150
              example: the water is clear
            Photos:
              type: array
              items:
                type: string
              minItems: 1
              maxItems: 3
        assignedMaintainerId:
          type: integer
          nullable: true
        followerUsers:
          type: array
          items:
            $ref: "#/components/schemas/User"
...
    Notification:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        reportId:
          type: integer
        type:
          type: string
          enum:
            - STATUS_CHANGE
            - OFFICER_MESSAGE
        message:
          type: string
        createdAt:
          type: string
          format: date-time
        read:
          type: boolean
...
    InternalMessage:
      type: object
      properties:
        id:
          type: integer
        reportId:
          type: integer
        senderType:
            $ref: "#/components/schemas/OfficerRole"
        senderId:
            type: integer
        receiverType:
            $ref: "#/components/schemas/OfficerRole"
        receiverId:
            type: integer
        message:
          type: string
        createdAt:
          type: string
          format: date-time
    Follow:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        reportId:
          type: integer
        createdAt:
          type: string
          format: date-time
    Statistics:
      type: array
      items:
        type: object
        properties:
          date:
            type: string
          totalReports:
            type: integer
          approvedReports:
            type: integer
          rejectedReports:
            type: integer
...
    Error:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
      required:
        - code
        - message
...
  securitySchemes:
    petstore_auth:
      type: oauth2
      flows:
        implicit:
          authorizationUrl: https://petstore3.swagger.io/oauth/authorize
          scopes:
            write:pets: modify pets in your account
            read:pets: read your pets
    api_key:
      type: apiKey
      name: api_key
      in: header